Add google cpplint test, enforcing a subset of rules
Also change all includes from "fossa.h" to "../fossa.h"
so that tools like flycheck don't need special flags to find
the headers.

Fixes all style violations and excludes sha1.c which is foreign code.
Line length set to 80 cols, following the rules in the developer primer doc.

The cpplint is not yet enabled in the default tests, to minimize
dependencies needed by a random developer. It can be enabled in the CI build.