Add 'local' includes to headers
Local headers need to be checkable for syntax in isolation
(with emacs flycheck or equivalent). This includes are then stripped away
in the amalgamated version like we did for `.c` files.
This can also be used in future to generate the correct order of amalgamation
and detect dependency cycles.

Also, fixes docs makefile to also depend on .h files,
useful now since we parse structs.